Add some stuff to protect against 24V reverse polarity
The charge pump section of the VNH5019 can only handle -16V, so would be
destroyed by -24V applied to the board. The new bit of circuitry ensures
that the reverse polarity FET is off when reverse polarity is applied,
while ensuring that no negative voltage is applied to the VNH5019 charge
pump at all.