Remove superfluous diode
The body diode in the reverse polarity protection FET (Q2) already
performs the function of the removed diode. The diode was only there to
provide power to the VBATT pins of the VNH5019s. We don't actually need
the functionality of the VBATT pins any more (their function is to
handle large negative voltages without damage). To ensure we're
operating the VNH5019s within spec the VBATT pins should still be
powered, but there should be no problem just powering them from the same
rail as the VCC lines.