Remove the footprint name from the description field
PCB doesn't have the ability to not output a bit of text for each
footprint while still outputting the rest of the silk. It does allow you
to select between three different things, so I've set one of these to be
empty for all footprints and told it to output that.

The board is too dense to fit the reference designators in. I will place
them in the middle of the components for as assembly aid.