Alter how the isolation bypass is performed
This is quite horrible and hacky at the moment, but the bypass pads are
included in the SOIC8 footprint itself. This means that the symbols
shown on the schematic are not functional and do not have footprints in
the layout. Unfortunately I can't find a way of telling PCB this, so
I've just disconnected them in the schematic for now.