Switch to a bigger uC
The R6 and R8 are pin compatible. The R8 has 64k of flash and 8k of RAM.
The main reason for switching is that the R8 has TIM4, which means that
both TIM3 and TIM4 will be exposed on the expansion header. This will
allow for really easy quadrature encoder based feedback for both
channels. Even if we don't take advantage of this straight away, it'll
be really annoying in the future if we can't do it because of a weedy
micro.